Add video record to README
Added a link to a screencast in the README.

fixing video record issue

Fix video link

Updated the output section to correct the link and numbering.

Add files via upload

Update README.md

Fix output link in README.md

Updated the output link in the README file.

Add GitHub asset link to README

Added a link to the GitHub asset in the README.